\name{\[}

\alias{\[}
\alias{\[,MoveStack,ANY,ANY-method}

\docType{methods}

\title{Returns an object from a MoveStack}

\description{Returning a single Move object}

\usage{
  \S4method{\[}{MoveStack,ANY,ANY}(x,i,y,drop)
}

\arguments{
  \item{x}{...}
  \item{i}{...}
  \item{j}{...}
  \item{drop}{...}
}

\details{
  ...
}

\author{Marco Smolla}

\note{
  ...
}
