\name{as.data.frame}

\alias{as.data.frame,Move-method}

\docType{methods}

\title{Return a Data Frame}

\description{Functions to create a data.frame with the spatial data frame information from a Move object.}

\usage{
  \S4method{as.data.frame}{Move}(x,...)
}

\arguments{
  \item{x}{a Move object}
  \item{...}{additional arguments to be passed to or from methods}
}

\details{
  as.data.frame extracts the sdf argument from a Move object
}


\author{Marco Smolla}
