% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{clamp}
\alias{clamp}
\title{Clamp values to a specified range}
\usage{
clamp(x, min = -1, max = 1)
}
\arguments{
\item{x}{Numeric vector.}

\item{min, max}{Numeric scalars specifying the lower and upper limits.}
}
\value{
A numeric vector where all values lie within \verb{[min, max]}.
}
\description{
Restricts numeric values to lie between given minimum and maximum bounds.
}
\examples{
clamp(c(-2, 0.5, 3), 0, 1)
}
