\name{dBGBvariance-class}
\alias{dBGBvariance-class}
\alias{dBGBvarianceTmp-class}
%\alias{dynBGBvariance}
%\alias{dynBGBvarianceTmp}
\docType{class}
\title{Class to store the orthogonal and parallel variance}
\description{
	This class stores the orthogonal and parellel variances calculated with the \code{dynBGBvariance} function.
}

