\name{dBMvarianceTmp}

\alias{dBMvarianceTmp-class}

\docType{class}

\title{dBMvarianceTmp class}

\description{The class dBMvarianceTmp}

\section{Slots}{
  \describe{
    \item{window.size}{...}
    \item{margin}{...}
    \item{means}{...}
    \item{in.windows}{...}
    \item{interest}{...}
    \item{break.list}{...}    
  }
}

\author{Marco Smolla}