% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{degrees_to_radians}
\alias{degrees_to_radians}
\title{Convert degrees to radians}
\usage{
degrees_to_radians(degrees)
}
\arguments{
\item{degrees}{Numeric scalar or vector of angles in degrees.}
}
\value{
Numeric value(s) in radians.
}
\description{
Converts an angle expressed in degrees to its equivalent value in radians.
}
\examples{
degrees_to_radians(180)  # pi
degrees_to_radians(90)   # pi/2

}
\seealso{
\code{\link[=radians_to_degrees]{radians_to_degrees()}}
}
