\docType{methods}
\name{dynBGB}
\alias{dynBGB}
\title{Calculation of the dynamic bivariate gausian bridge}
\arguments{
  \item{move}{the move object or variance object used for
  calculating the ud if a \code{.MoveTrackSingle} object is
  supplied this is converted into a \code{dBGBvariance}
  object using the \code{dynBGBvariance} function}

  \item{raster}{either the raster used for variance
  calculation or the number of sells the raster need to
  consist of}

  \item{locErr}{the location errors used for the
  calculation}

  \item{maxInt}{}
}
\description{
}

