\name{dynBGBvariance}
\alias{dynBGBvariance}
\alias{dynBGBvariance,.MoveTrackSingle,numeric,numeric,numeric-method}
\docType{methods}
\title{calculate variance for a track using a running window}
\description{
  the function uses \code{windowApply} with the
  \code{BGBvarbreak} function in order to implement a
  dynaminc calculation of the variance
}

