\name{getMovebankStudy}

\alias{getMovebankStudy}
\alias{getMovebankStudy,ANY,missing-method}
\alias{getMovebankStudy,numeric,CURLHandle-method}
\alias{getMovebankStudy,character,CURLHandle-method}

\docType{methods}

\title{Returns information about the requested study}

\description{This function returns information about the requested study.}

\usage{
  getMovebankStudy(study, login, ...)
}

\arguments{
  \item{study}{a character string (study name) or the numeric study ID as it is stored on Movebank}
  \item{login}{...}
  \item{...}{...}
}

\details{
  ...
}


\author{Marco Smolla}