\name{lineMidpoint}

\alias{lineMidpoint}
\alias{lineMidpoint,SpatialPointsDataFrame-method}

\docType{methods}

\title{Calculates Line mid points ...}

\description{Points in the middle of a line ...}

\usage{
  \S4method{lineMidpoint}{SpatialPointsDataFrame}(object)
}

\arguments{
  \item{object}{...}
}

\details{
  ...
}

\author{Marco Smolla}

\note{
  ...
}

\examples{
  \dontrun{
    
  }
}