% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{magnitude}
\alias{magnitude}
\title{Compute the magnitude (Euclidean norm) of a vector}
\usage{
magnitude(x)
}
\arguments{
\item{x}{Numeric vector.}
}
\value{
Numeric scalar representing the magnitude of \code{x}.
}
\description{
Calculates the length of a numeric vector using the Euclidean (L2) norm.
}
\examples{
magnitude(c(3, 4))  # returns 5
}
