% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{measure_angle_between_planes}
\alias{measure_angle_between_planes}
\title{Compute angle between two planes}
\usage{
measure_angle_between_planes(n1, n2, degrees = TRUE)
}
\arguments{
\item{n1, n2}{Numeric vectors (length 3) giving the normal vectors of the planes.}

\item{degrees}{Logical; if \code{TRUE}, return the angle in degrees (default).}
}
\value{
A numeric scalar giving the unsigned angle between the two planes.
}
\description{
Calculates the (unsigned) smallest angle between two planes given their
normal vectors. Returns an angle in degrees by default.
}
\details{
The returned value is always in the range \eqn{[0, 180]} and does not encode
rotation direction. For an oriented (signed) version, see
\code{\link{measure_signed_angle_between_planes}}.
}
\examples{
n1 <- c(0, 0, 1)
n2 <- c(0, 1, 0)
measure_angle_between_planes(n1, n2)

}
\seealso{
\code{\link[=measure_signed_angle_between_planes]{measure_signed_angle_between_planes()}}
}
