\name{MovebankLogin-class}

\alias{MovebankLogin}
\alias{MovebankLogin-class}

\docType{class}

\title{
  The MovebankLogin class
}

\description{
  The MovebankLogin object is needed for every Movebank browsing function. Alternatively, one can also chose to enter the username and password every time one uses one of the browsing functions. The object is inhereted from an \code{httr} \code{request} obnject.
}



\section{Methods}{
  \describe{
    \item{\code{\link{movebankLogin}}}{\code{signature(object = "character")}: creates a MovebankLogin object}}
}

\author{Bart Kranstauber \& Marco Smolla}

\keyword{classes}