\name{MoveBurst}

\alias{MoveBurst-class}

\docType{class}

\title{MoveBurst class}

\description{The class MoveBurst}

\section{Slots}{
  \describe{
    \item{bbox}{...}
    \item{burstId}{...}
    \item{citation}{...}
    \item{coords}{...}
    \item{coords.nrs}{...}
    \item{data}{...}
    \item{dateCreation}{...}
    \item{idData}{...}
    \item{license}{...}
    \item{proj4string}{...}
    \item{sensor}{...}
    \item{study}{...}
    \item{timesMissedFixes}{...}
    \item{timestamps}{...}
  }
}

\author{Marco Smolla}