\name{n.locs}

\alias{n.locs}
\alias{n.locs,Move-method}

\docType{methods}

\title{Extract the number of locations of a track from a Move object}

\description{The n.locs method returns the number of locations of a track from a Move object.}

\usage{
  \S4method{n.locs}{Move}(obj)
}

\arguments{
  \item{obj}{A Move object}
}

\details{
  \code{obj} a Move object
}

\author{Marco Smolla}