% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{normalise}
\alias{normalise}
\title{Normalize a numeric vector}
\usage{
normalise(x)
}
\arguments{
\item{x}{Numeric vector.}
}
\value{
A numeric vector of the same length as \code{x}, scaled so that its magnitude is 1.
}
\description{
Scales a vector to unit length by dividing each element by its Euclidean norm.
If input is a zero vector, returns vector of NaN values
}
\examples{
normalise(c(3, 4))  # returns c(0.6, 0.8)
}
