% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{radians_to_degrees}
\alias{radians_to_degrees}
\title{Convert radians to degrees}
\usage{
radians_to_degrees(radians)
}
\arguments{
\item{radians}{Numeric scalar or vector of angles in radians.}
}
\value{
Numeric value(s) in degrees.
}
\description{
Converts an angle expressed in radians to its equivalent value in degrees.
}
\examples{
radians_to_degrees(pi)      # 180
radians_to_degrees(pi / 2)  # 90

}
\seealso{
\code{\link[=degrees_to_radians]{degrees_to_radians()}}
}
