\name{raster}

\alias{raster,DBBMM-method}

\docType{methods}

\title{Extract raster from DBBMM}

\description{raster plot of a RasterLayer from a DBBMM object.}

\usage{
  \S4method{raster}{DBBMM}(x)
}

\arguments{
  \item{x}{a DBBMM object}
}

\details{
  The raster function extraxcts the raster object from the DBBMM object. It returns an object from class RasterLayer.
}

\author{Marco Smolla}