% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{rotate_vector_into_a_plane}
\alias{rotate_vector_into_a_plane}
\title{Compute the rotation needed to align a 3D vector with a plane}
\usage{
rotate_vector_into_a_plane(
  v,
  plane_normal,
  return = c("v_new", "axis_plus_angle")
)
}
\arguments{
\item{v}{numeric(3) Vector to rotate.}

\item{plane_normal}{numeric(3) Plane normal (not necessarily unit length).}

\item{return}{whether to return the new (transformed) vector or a list of paramaters required to perform the rotation (axis and angle).}
}
\value{
either a rotated vector or a list with elements \code{axis} (unit vector) and \code{angle} (radians) depending on value of \code{return} param.
}
\description{
Computes the minimal rotation required to move a 3D vector \code{v} so that it lies
within the plane defined by its normal vector \code{plane_normal}. The rotation
preserves the magnitude of \code{v}, unlike a projection which shortens it.
}
\examples{
rotate_vector_into_a_plane(c(1,2,3), c(0,1,-1))
}
