\name{split}

\alias{split}
\alias{split,MoveStack,missing-method}

\docType{methods}

\title{Splitting MoveStack}

\description{Splitting MoveStack in a list of Move objects}

\usage{
  \S4method{split}{MoveStack,missing}(x, f, drop=FALSE, ...)
}

\arguments{
  \item{x}{MoveStack object}
  \item{f}{not needed}
  \item{drop}{not needed}
  \item{...}{not needed}
}

\details{
  ...
}

\author{Marco Smolla}

\note{
  ...
}

\examples{
sum(2,1)
}