\name{time.lag}

\alias{time.lag}
\alias{time.lag,Move-method}

\docType{methods}

\title{Calculates the time lags between the coordinates}

\description{The time.lag function calculates the time lags between coordinates.}

\usage{
  \S4method{time.lag}{Move}(x,...)
}

\arguments{
  \item{x}{A Move object}
  \item{...}{Further arguments ...}
}

\details{
  \code{...} there are no further arguments
}

\author{Marco Smolla}
