% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{translate_table_in_direction}
\alias{translate_table_in_direction}
\title{Translate all points in a table in direction}
\usage{
translate_table_in_direction(table, direction, magnitude, careful = TRUE)
}
\arguments{
\item{table}{A data frame or matrix containing at least the columns
\code{x}, \code{y}, and \code{z}.}

\item{direction}{Numeric vector giving the direction of translation (need not be unit length).}

\item{magnitude}{Numeric scalar giving the distance to move along \code{direction}.}

\item{careful}{Should we assert the input is sensible? (slow but safer)}
}
\value{
The input \code{table} translated along \code{direction} by \code{magnitude}
}
\description{
Rotates each row of a coordinate table (with columns \code{x}, \code{y}, \code{z}) about a
specified axis by a given angle. The axis may pass through an arbitrary point
(default is the origin).
}
\examples{
pts <- data.frame(x = c(2, 3), y = c(0, 0), z = c(0, 0))
direction <- c(0, 0, 1)
magnitude <- 10
translate_table_in_direction(pts, direction, magnitude)

}
