\name{utilization density data}
\alias{leroydbbmm}
\alias{dbbmmstack}

\docType{data}
\title{
Dynamic brownian bridges
}
\description{
Utilization densities calculated with \code{brownian.bridge.dyn} to examplify functions.
}
\usage{data("leroydbbmm")
}

\details{
see createRDataFile.R in inst/extdata for the exact calculation
}


\examples{
data(dbbmmstack)
data(leroydbbmm)
leroydbbmm
}
\keyword{datasets}
