% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{wrap_to_2pi}
\alias{wrap_to_2pi}
\title{Wrap angles to [0, 2*pi) radians}
\usage{
wrap_to_2pi(radians)
}
\arguments{
\item{radians}{Numeric vector of angles in radians.}
}
\value{
Numeric vector of angles wrapped to [0, 2*pi).
}
\description{
Wraps input angles so that the result lies within the range [0, 2*pi).
}
\examples{
wrap_to_2pi(c(-pi/2, 0, 5*pi))
# Returns approximately c(4.712389, 0, 3.141593)
}
