% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{wrap_to_360}
\alias{wrap_to_360}
\title{Wrap angles to [0, 360) degrees}
\usage{
wrap_to_360(degrees)
}
\arguments{
\item{degrees}{Numeric vector of angles in degrees.}
}
\value{
Numeric vector of angles wrapped to [0, 360).
}
\description{
Wraps input angles so that the result lies within the range [0, 360).
}
\examples{
wrap_to_360(c(-30, 0, 370))
# Returns c(330, 0, 10)
}
