\name{burstTrack}

\alias{burstTrack}
\alias{burstTrack,Move,integer-method}

\docType{methods}

\title{Bursting a track}

\description{Bursting a track by specified variable}

\usage{
  \S4method{burstTrack}{Move,integer}(object, by, breaks=5, sizeFUN="relTime")
}

\arguments{
  \item{object}{...}
  \item{by}{...}
  \item{breaks}{...}
  \item{sizeFUN}{...}
  \item{...}{additional plot attributes}
}

\details{
  ...
}

\author{Marco Smolla}

\note{
  ...
}

\examples{
  \dontrun{
    
  }
}