\name{dBMvarianceTmp}

\alias{dBMvarianceTmp-class}

\docType{class}

\title{dBMvarianceTmp class}

\description{The class dBMvarianceTmp is mostly an internal class that is made public to make inheritance easier. It is a basal class that stores results of the dbbmm}

\section{Slots}{
  \describe{
    \item{window.size}{The window size used for dbbmm calculation}
    \item{margin}{The margin used for dbbmm calculation}
    \item{means}{...}
    \item{in.windows}{...}
    \item{interest}{...}
    \item{break.list}{...}    
  }
}

\author{Marco Smolla}
