\name{getMovebankSensorsAttributes}

\alias{getMovebankSensorsAttributes}
\alias{getMovebankSensorsAttributes,character,MovebankLogin-method}
\alias{getMovebankSensorsAttributes,numeric,MovebankLogin-method}

\docType{methods}

\title{Returns additional sensor attributes}

\description{This function returns all additional attributes of the sensors of the requested study.}

\usage{
  getMovebankSensorsAttributes(study, login)
}

\arguments{
  \item{study}{a character string (study name) or the numeric study ID as it is stored on Movebank}
  \item{login}{...}
}

\details{
  ... 
}


\author{Marco Smolla}