\name{MovebankLogin-class}

\alias{MovebankLogin}
\alias{MovebankLogin-class}

\docType{class}

\title{
  The MovebankLogin class
}

\description{
  The MovebankLogin object is from every Movebank browsing function and created by movebankLogin
}

\section{Slots}{
  \describe{
    \item{username}{Object of class \code{"dBMvariance"}: includes the break.list and points of interest, ...; }
    \item{password}{  }
    \item{rcurl}{the extension factor set by the user}

  }
}

\section{Methods}{
  \describe{
    \item{\code{\link{movebankLogin}}}{\code{signature(object = "MovebankLogin")}: adds a contour line to a plot}}
}

\author{Marco Smolla}

\keyword{classes}