\name{MoveStack-class}

\alias{MoveStack}
\alias{MoveStack-class}

\docType{class}

\title{
  The MoveStack class
}

\description{
  The MoveStack object is created within the brownian.bridge.dyn function from a Move object. It includes beneath others a raster object and probabilities.
}

\section{Slots}{
  \describe{
    \item{citation}{Object of class \code{"dBMvariance"}: includes the break.list and points of interest, ...; }
    \item{bbox}{  }
    \item{coords}{the extension factor set by the user}
    \item{coords.nrs}{  }
    \item{data}{  }
    \item{dateCreation}{  }
    \item{idData}{  }
    \item{license}{  }
    \item{proj4string}{  }
    \item{study}{  }
    \item{timestamps}{  }
    \item{trackId}{  }
  }
}

\author{Marco Smolla}

\keyword{classes}