\name{moveStack}

\alias{moveStack}
\alias{moveStack,list-method}
\alias{moveStack,character-method}

\docType{methods}

\title{Creating a MoveStack}

\description{Either stacks a list of Move objects or creates a MoveStack from a csv file}

\usage{
  \S4method{moveStack}{list}(x)
  \S4method{moveStack}{character}(x, proj)
}

\arguments{
  \item{x}{...}
  \item{proj}{...}
}

\details{
  ...
}

\author{Marco Smolla}

\note{
  ...
}

\examples{
  \dontrun{
    
  }
}