\name{plotBursts}

\alias{plotBursts}
\alias{plotBursts,list-method}
\alias{plotBursts,SpatialPointsDataFrame-method}

\docType{methods}

\title{Plotting the centroids of a track}

\description{Plotting centroids ...}

\usage{
  \S4method{plotBursts}{list}(object, add=TRUE, ...)
  \S4method{plotBursts}{SpatialPointsDataFrame}(object, add, ...)
}

\arguments{
  \item{object}{...}
  \item{add}{...}
  \item{...}{additional plot attributes}
}

\details{
...
}

\author{Marco Smolla}

\note{
...
}

\examples{
  \dontrun{

  }
}