\name{raster2contour}

\alias{raster2contour}
\alias{raster2contour,RasterLayer-method}

\docType{methods}

\title{Converting raster to contour lines}

\description{Plotting centroids ...}

\usage{
  \S4method{raster2contour}{RasterLayer}(x, ...)
}

\arguments{
  \item{x}{...}
  \item{...}{additional plot attributes}
}

\details{
  ...
}

\author{Marco Smolla}

\note{
  ...
}

\examples{
  \dontrun{
    
  }
}