\name{subset-method}

\alias{[,MoveStack,ANY,ANY-method}
\alias{[,.MoveTrack,ANY,ANY-method}
\alias{[,.MoveTrackStack,ANY,ANY-method}
\alias{[,dBMvariance,ANY,ANY-method}
\alias{[,dBMvarianceStack,ANY,ANY-method}

\docType{methods}

\title{Returns a single object from a MoveStack}

\description{Returning a single Move object}

\usage{
  \S4method{[}{MoveStack,ANY,ANY}(x,i)
}

\arguments{
  \item{x}{MoveStack object}
  \item{i}{numeric, number of position in the stack}
}


\author{Marco Smolla}
