\name{subset-method}

\alias{[,MoveStack,ANY,ANY-method}

\docType{methods}

\title{Returns an object from a MoveStack}

\description{Returning a single Move object}

\usage{
  \S4method{[}{MoveStack,ANY,ANY}(x,i)
}

\arguments{
  \item{x}{...}
  \item{i}{...}
}

\details{
  ...
}

\author{Marco Smolla}

\note{
  ...
}
