\name{summary}

\alias{summary}
\alias{summary,DBBMM-method}

\docType{methods}

\title{A summary of a DBBMM object}

\description{Summarizes the information of the raster from a DBBMM object}

\usage{
  \S4method{summary}{DBBMM}(object)
}

\arguments{
  \item{object}{a DBBMM object}
}

\details{
Returns the projection, extent, and maximum and minimum values of the raster stored within the DBBMM object.
}

\author{Marco Smolla}