\name{summary}

\alias{summary}
\alias{summary,Move-method}

\docType{methods}

\title{Summarizes a Move object}

\description{The summary method returns a short form of the spatial data frame and the number of omitted locations.}

\usage{
  \S4method{summary}{Move}(object)
}

\arguments{
  \item{object}{A Move object}
}

\details{
  \code{object} a Move object
}

\author{Marco Smolla}