\name{time.lag}

\alias{time.lag}
\alias{time.lag,Move-method}

\docType{methods}

\title{Calculates the time lags between the coordinates}

\description{The time.lag function calculates the time lags between locations.}

\usage{
  \S4method{time.lag}{Move}(x,...)
}

\arguments{
  \item{x}{a Move object}
  \item{...}{further arguments}
}

\value{
  The function returns time lags in minutes
}

\author{Marco Smolla}
